requirements.txt
langchain
openai
pypdf
markdown2
scikit-learn
pydantic==1.10.13
tiktoken
python-dotenv
bs4
html2text
langchain_openai
langchain-community
uvicorn
pinecone-client
fastapi
langsmith
langchain-core
langserve
langchain-pinecone
langchain-voyageai
sse_starlette
pymongo
langchain-mongodb
hypercorn
apscheduler
python-jose
python-multipart
bcrypt
passlib
pyjwt
markdown 
passlib
sendgrid
youtube-transcript-api
pytube
pypdf2
python-docx